requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Broker
ib_insync>=0.9.86
pytz>=2023.3

# Market calendar (holiday checking)
pandas_market_calendars>=4.3.0

# Dashboard
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Notebooks
jupyter>=1.0.0
